requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.1
djangorestframework==3.12.1
gunicorn==20.0.4
httpie==2.2.0
idna==2.10
JSON-minify==0.3.0
lxml==4.5.2
psycopg2-binary==2.8.6
Pygments==2.7.1
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
